requirements.txt
django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
black==24.10.0
psycopg[binary]
psycopg2
coverage==7.6.2


